fix JWT_SECRET not using env, add queue job, cleanup debug statements

+++ src/jobs/queue.js
@@ -1,5 +1,5 @@
 import d from 'debug';
-import tumblr from 'tumblr.js';
+import Tumblr from 'tumblr.js';
 import config from '../config';
 import {User, Post, Queue, TumblrAccount} from '../models';
 
@@ -11,12 +11,21 @@ export default agenda => {
         const {queueId, userId} = job.attrs.data;
         debug(`Checking if queue ${queueId} is allowed to post`);
 
-        const user = await User.findOne({_id: userId}).select('suspended').lean().exec().catch(done);
+        const user = await User.findOne({_id: userId}).select('suspended').lean().exec().catch(err => {
+            debug(err);
+            debug('calling done');
+            return done(err);
+        });
+
         if (user.suspended) {
             return done(`Cannot check queue as user ${userId} is suspended`);
         }
 
-        const queue = await Queue.findOne({_id: queueId}).select('startTime endTime enabled').lean().exec().catch(done);
+        const queue = await Queue.findOne({_id: queueId}).select('startTime endTime enabled').lean().exec().catch(err => {
+            debug(err);
+            debug('calling done');
+            return done(err);
+        });
 
         const now = (new Date()).getTime();
         const midnight = (new Date()).setHours(0, 0, 0, 0);
@@ -31,44 +40,84 @@ export default agenda => {
             return done('Cannot schedule queue process as queue is disabled');
         }
 
-        agenda.now('process queue', {
-            queueId: queue._id,
-            userId: user._id
+        const newJob = agenda.create('process queue', {userId: user._id, queueId: queue._id});
+        newJob.unique({'data.userId': user._id, 'data.queueId': queue._id});
+        newJob.schedule(new Date());
+        newJob.save(err => {
+            if (err) {
+                debug(err);
+                debug('calling done');
+                return done(err);
+            }
         });
 
+        debug(`Scheduling queue [${queue._id}] to be processed`);
         return done();
     });
 
     agenda.define('process queue', async (job, done) => {
-        const {queueId, userId} = job.attrs.data;
-        debug(`Trying to process queue ${queueId} for ${userId}`);
-        const queue = await Queue.findOne({_id: queueId}).populate('blogs', '_id url').lean().exec().catch(done);
+        const {userId, queueId} = job.attrs.data;
+        debug(`Trying to process queue [%s] for user [%s]`, queueId, userId);
+        const queue = await Queue.findOne({_id: queueId}).populate('blogs', '_id url').populate('posts').exec().catch(err => {
+            debug(err);
+            debug('calling done');
+            return done(err);
+        });
+
+        debug('Queue returned %O', queue);
+
+        if (queue.posts.length === 0) {
+            return done(`Cannot process queue as it's empty`);
+        }
 
         // First post
         const [post] = queue.posts;
+        debug('Trying to process post [%s] to queue [%s]', post._id, queue._id);
 
         queue.blogs.forEach(async blog => {
-            const {token, tokenSecret} = await TumblrAccount.findOne({blogs: blog._id}).lean().exec().catch(done);
-            const client = tumblr.createClient({
-                consumer_key: process.env.TUMBLR_TOKEN || config.get('tumblr.token'), // eslint-disable-line camelcase
-                consumer_secret: process.env.TUMBLR_TOKEN_SECRET || config.get('tumblr.tokenSecret'), // eslint-disable-line camelcase
-                token,
-                token_secret: tokenSecret // eslint-disable-line camelcase
+            debug(`Processing blog [%s]`, blog._id);
+            const {_id, token, secret} = await TumblrAccount.findOne({blogs: blog._id}).select('token secret').lean().exec().catch(err => {
+                debug(err);
+                debug('calling done');
+                return done(err);
             });
 
-            client.reblog(blog.url, {
+            if (!token || !secret) {
+                debug(`TumblrAccount [${_id}] is missing it's token or secret`);
+                return done(`TumblrAccount [${_id}] is missing it's token or secret`);
+            }
+
+            const client = new Tumblr.Client({
+                credentials: {
+                    consumer_key: process.env.TUMBLR_TOKEN || config.get('tumblr.token'), // eslint-disable-line camelcase
+                    consumer_secret: process.env.TUMBLR_TOKEN_SECRET || config.get('tumblr.tokenSecret'), // eslint-disable-line camelcase
+                    token,
+                    token_secret: secret // eslint-disable-line camelcase
+                }
+            });
+
+            debug(`Successfully connected to Tumblr for blog [%s]`, blog._id);
+
+            client.reblogPost(blog.url, {
                 id: post.postId,
                 reblog_key: post.reblogKey // eslint-disable-line camelcase
-            }, async err => {
+            }, async (err, resp) => {
                 if (err) {
+                    debug(err);
                     return done(err);
                 }
+
+                debug(`Tumblr returned %O`, resp);
+
+                debug(`Successfully reblogged [%s] to Tumblr for blog [%s]`, post.postId, blog._id);
 
                 try {
                     await Post.findOneAndRemove({_id: post._id}).lean().exec();
                     queue.posts = queue.posts.splice(queue.posts.findIndex(x => post._id === x._id), 1);
                     await queue.save();
+                    debug('Finished processing queue [%s]', queue._id);
                 } catch (err) {
+                    debug(err);
                     return done(err);
                 }
             });
